Add support for Folsom's python-glanceclient
Fixes #20

Heat supports both Essex and Folsom for now. However, in Folsom, the use
of python-glance client is deprecated in favor of python-glanceclient.

Since the clients are incompatible, this adds the `heat.glance_client`
package that abstracts the differences away.

Once we stop supporting Essex, most of this should go away.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>