Drop lower-constraints.txt and its testing
As discussed in TC PTG[1] and TC resolution[2], we are
dropping the lower-constraints.txt file and its testing.
We will keep lower bounds in the requirements.txt file but
with a note that these are not tested lower bounds and we
try our best to keep them updated.

[1] https://etherpad.opendev.org/p/tc-zed-ptg#L326
[2] https://governance.openstack.org/tc/resolutions/20220414-drop-lower-constraints.html#proposal

This also makes the following two changes to fix the broken CI.

- Switch the python3 job template to unversioned one because
  the heat-cfntools project follows an independent release model.
  This effectively removes py36 tests.

- Use upper-constraints.txt to build documentation, to avoid pulling
  the latest Sphinx which includes a breaking change.

Change-Id: I3c41ed7cd54e8c8d3fad05ea7ac86fc35b9fa4d1